backend/requirements-render.txt
# Kidney Ultrasound Classification Backend
# Production requirements for Render deployment

# ============================================================================
# Core Web Framework
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ============================================================================
# Lightweight Inference Engine (Memory-Optimized for Free Tier)
# ============================================================================
# ONNX Runtime: 50MB - lightweight CPU inference
onnxruntime>=1.16.0

# ============================================================================
# PyTorch (ONE-TIME USE: .pth to .onnx conversion only)
# ============================================================================
# Only used on first deployment to convert .pth model to .onnx format.
# After conversion, torch can be uninstalled (frees 1+ GB).
# Re-add to requirements if you need to regenerate ONNX models.
# torch>=2.0.0,<3.0.0
# torchvision>=0.15.0,<1.0.0

# ============================================================================
# Image Processing
# ============================================================================
opencv-python-headless>=4.8.0  # For headless environments (Render)
pillow>=10.0.0

# ============================================================================
# Data Science & Numerical
# ============================================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ============================================================================
# Configuration & Logging
# ============================================================================
pyyaml>=6.0
loguru>=0.7.0

# ============================================================================
# Utilities
# ============================================================================
requests>=2.31.0  # For downloading models
matplotlib>=3.7.0

# ============================================================================
# Production Server (optional, for better performance)
# ============================================================================
gunicorn>=21.2.0  # Optional: better performance than uvicorn alone

# ============================================================================
# Monitoring & Health (optional)
# ============================================================================
# prometheus-client>=0.17.0  # Uncomment for metrics
